Move tokenWarning to its own config
