+++ apps/ai-content-generator/src/configs/features/featureTypes.tsx
@@ -10,18 +10,11 @@ export type Prompt = (input: string, targetLocale: string, ...args: string[]) =>
  * Feature Items
  */
 
-export type TokenWarning = {
-  warningText: string;
-  substring: string;
-  link: string;
-};
-
 export type DialogText = {
   fieldHelpText: string;
   fieldPlaceholder: string;
   promptHelpText: string;
   promptPlaceholder: string;
-  tokenWarning: TokenWarning;
 };
 
 type FeatureItem = {
